requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
customtkinter==5.2.1
cycler==0.12.1
darkdetect==0.8.0
fonttools==4.47.0
idna==3.6
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.2
packaging==23.2
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
pywinstyles==1.4
requests==2.31.0
six==1.16.0
urllib3==2.1.0
